Data.Vector.Circular

Types

data CircularVector a

A circular, immutable vector. This type is equivalent to cycle xs for some finite, nonempty xs, but with O(1) access and O(1) rotations. Indexing into this type is always total.

Constructors

CircularVector

Fields

vector :: !(NonEmptyVector a)

rotation :: !Int

Construction

singleton :: a -> CircularVector a

Construct a singleton 'CircularVector.

Since: 0.1

toVector :: CircularVector a -> Vector a

Construct a Vector from a CircularVector.

Since: 0.1

toNonEmptyVector :: CircularVector a -> NonEmptyVector a

Construct a NonEmptyVector from a CircularVector.

Since: 0.1.1

fromVector :: NonEmptyVector a -> CircularVector a

Construct a CircularVector from a NonEmptyVector.

Since: 0.1

unsafeFromVector :: Vector a -> CircularVector a

Construct a CircularVector from a Vector.

Calls error if the input vector is empty.

Since: 0.1

fromList :: [a] -> Maybe (CircularVector a)

Construct a CircularVector from a list.

Since: 0.1

fromListN :: Int -> [a] -> Maybe (CircularVector a)

Construct a CircularVector from a list with a size hint.

Since: 0.1

unsafeFromList :: [a] -> CircularVector a

Construct a CircularVector from a list.

Calls error if the input list is empty.

Since: 0.1

unsafeFromListN :: Int -> [a] -> CircularVector a

Construct a CircularVector from a list with a size hint.

Calls error if the input list is empty, or if the size hint is <= 0.

Since: 0.1

vec :: Lift a => [a] -> Q (TExp (CircularVector a))

Construct a CircularVector at compile-time using typed Template Haskell.

Since: 0.1

Rotation

rotateLeft :: Int -> CircularVector a -> CircularVector a

Rotate the vector to the left by n number of elements.

Note: Left rotations start to break down due to arithmetic underflow when the size of the input vector is > maxBound Int@

Since: 0.1

rotateRight :: Int -> CircularVector a -> CircularVector a

Rotate the vector to left by n number of elements.

Note: Right rotations start to break down due to arithmetic overflow when the size of the input vector is > maxBound Int@

Since: 0.1

Comparisons

equivalent :: Ord a => CircularVector a -> CircularVector a -> Bool

Since: 0.1

canonise :: Ord a => CircularVector a -> CircularVector a

Since: 0.1

leastRotation :: forall a. Ord a => Vector a -> Int

Since: 0.1

Folds

foldMap :: Monoid m => (a -> m) -> CircularVector a -> m

Lazily-accumulating monoidal fold over a CircularVector. @since 0.1

foldMap' :: Monoid m => (a -> m) -> CircularVector a -> m

Strictly-accumulating monoidal fold over a CircularVector.

Since: 0.1

foldr :: (a -> b -> b) -> b -> CircularVector a -> b

Since: 0.1

foldl :: (b -> a -> b) -> b -> CircularVector a -> b

Since: 0.1

foldr' :: (a -> b -> b) -> b -> CircularVector a -> b

Since: 0.1

foldl' :: (b -> a -> b) -> b -> CircularVector a -> b

Since: 0.1

foldr1 :: (a -> a -> a) -> CircularVector a -> a

Since: 0.1

foldl1 :: (a -> a -> a) -> CircularVector a -> a

Since: 0.1

foldMap1 :: Semigroup m => (a -> m) -> CircularVector a -> m

Lazily-accumulating semigroupoidal fold over a CircularVector.

Since: 0.1

foldMap1' :: Semigroup m => (a -> m) -> CircularVector a -> m

Strictly-accumulating semigroupoidal fold over a CircularVector.

Since: 0.1

toNonEmpty :: CircularVector a -> NonEmpty a

Since: 0.1

Specialized folds

all :: (a -> Bool) -> CircularVector a -> Bool

O(n) Check if all elements satisfy the predicate.

Since: 0.1.1

any :: (a -> Bool) -> CircularVector a -> Bool

O(n) Check if any element satisfies the predicate.

Since: 0.1.1

and :: CircularVector Bool -> Bool

O(n) Check if all elements are True.

Since: 0.1.1

or :: CircularVector Bool -> Bool

O(n) Check if any element is True.

Since: 0.1.1

sum :: Num a => CircularVector a -> a

O(n) Compute the sum of the elements.

Since: 0.1.1

product :: Num a => CircularVector a -> a

O(n) Compute the product of the elements.

Since: 0.1.1

maximum :: Ord a => CircularVector a -> a

O(n) Yield the maximum element of the circular vector.

Since: 0.1.1

maximumBy :: (a -> a -> Ordering) -> CircularVector a -> a

O(n) Yield the maximum element of a circular vector according to the given comparison function.

Since: 0.1.1

minimum :: Ord a => CircularVector a -> a

O(n) Yield the minimum element of the circular vector.

Since: 0.1.1

minimumBy :: (a -> a -> Ordering) -> CircularVector a -> a

O(n) Yield the minimum element of a circular vector according to the given comparison function.

Since: 0.1.1

rotateToMinimumBy :: (a -> a -> Ordering) -> CircularVector a -> CircularVector a

O(n) Rotate to the minimum element of the circular vector according to the given comparison function.

Since: 0.1.1

rotateToMaximumBy :: (a -> a -> Ordering) -> CircularVector a -> CircularVector a

O(n) Rotate to the maximum element of the circular vector according to the given comparison function.

Since: 0.1.1

Indexing

index :: CircularVector a -> Int -> a

Index into a CircularVector. This is always total.

Since: 0.1

head :: CircularVector a -> a

Get the first element of a CircularVector. This is always total.

Since: 0.1

last :: CircularVector a -> a

Get the last element of a CircularVector. This is always total.

Since: 0.1

Zipping

zipWith :: (a -> b -> c) -> CircularVector a -> CircularVector b -> CircularVector c

O(min(m,n)) Zip two circular vectors with the given function.

Since: 0.1.1

zipWith3 :: (a -> b -> c -> d) -> CircularVector a -> CircularVector b -> CircularVector c -> CircularVector d

Zip three circular vectors with the given function.

Since: 0.1.1

zip :: CircularVector a -> CircularVector b -> CircularVector (a, b)

O(min(n,m)) Elementwise pairing of circular vector elements. This is a special case of zipWith where the function argument is (,)

Since: 0.1.1

zip3 :: CircularVector a -> CircularVector b -> CircularVector c -> CircularVector (a, b, c)

Zip together three circular vectors.

Since: 0.1.1

Permutations

reverse :: CircularVector a -> CircularVector a

O(n) Reverse a circular vector.

Since: 0.1.1